feat(WebWorkerProgressHandler): refactor and modularize progress handling
- Moved web worker progress handling logic to a new utility file for better organization and maintainability.
- Enhanced the initialization of progress notifications, ensuring proper tracking and cleanup of active worker tasks.
- Improved error handling and logging for task notifications, maintaining existing functionality while increasing robustness.
